added default blog setting, implemented preference service

+++ src/org/myjerry/evenstar/service/BlogPostService.java
@@ -20,7 +20,7 @@ public interface BlogPostService {
 	
 	public Collection<BlogPost> getAllBlogPosts(Long blogID);
 	
-	public Collection<BlogPost> getBlogPosts(Long blogID, long page, long count);
+	public Collection<BlogPost> getBlogPosts(Long blogID, int page, int count);
 	
 	public Long getTotalPosts(Long blogID);
 }
